zero-copy IPC batch ingestion
