+++ rust/sdk/Cargo.toml
@@ -29,6 +29,7 @@ tracing = "0.1.41"
 hyper-http-proxy = { version = "1.1", default-features = false }
 hyper-util = { version = "0.1", features = ["client-legacy", "tokio"] }
 smallvec = "1.15.1"
+bytes = "1"
 
 # Arrow Flight dependencies
 arrow-flight = { version = "56.2.0", default-features = false, optional = true }
